requirements.txt
playwright
python-dotenv
twocaptcha